Moved all the file managers into one file manager
